% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{AddToList}
\alias{AddToList}
\title{Title.}
\usage{
AddToList(filepath)
}
\arguments{
\item{filepath}{the filepath of output of MCL.}
}
\value{
A nested list.
}
\description{
\code{AddToList} add the text file from output of MCL to a nested list. output of MCL files will group
all the elements belong to one cluster in one line,seperated by whitespace. This fuction go through input
file line by line and put elements in same cluster in to a list.
}
\details{
Details.
}
\section{purpose}{
 preparation for create dataframe used for plotting.
}

\examples{
AddToList("MCL_example_output")
}
