% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{CreatDataFrame}
\alias{CreatDataFrame}
\title{Title.
\code{CreatDataFrame.R} creat dataframe used for plotting from nested list
Details.}
\usage{
CreatDataFrame(filepath)
}
\arguments{
\item{filepath}{filepath of the output of MCL}
}
\value{
a dataframe used for plotting.
}
\description{
Title.
\code{CreatDataFrame.R} creat dataframe used for plotting from nested list
Details.
}
\examples{
CreatDataFrame("MCL_example_output")
}
