% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runvismclApp}
\alias{runvismclApp}
\title{\code{runvismclApp} launch the shiny app distributed with this package framework}
\usage{
runvismclApp()
}
\description{
\code{runvismclApp} launches the shiny app for which the code has been placed in  \code{./inst/shiny-scripts/vismclApp/}.
}
