% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{vismcl}
\alias{vismcl}
\title{vismcl}
\usage{
vismcl(filepath, showName = FALSE, Clusternames = NULL,
  color = "grey")
}
\arguments{
\item{filepath}{the filepath of the output file of MCL}

\item{showName}{(optional)if shows the name of clusters,default is FALSE}

\item{Clusternames}{(optional) a vector for user-defined clusternames,
order should be the same as the lines go.}

\item{color}{(optional)A color code:color fill the circle,default is grey}
}
\value{
a circle packing graph representing the information
}
\description{
plot circle packing graph
}
\examples{
cat("a aa\\n b bbb b\\n",file = "A")
vismcl("A")
vismcl("A",showName = TRUE,color = "red")
Clust_names<-c("a","b")
vismcl("A",showName = TRUE,Clusternames = Clust_names)
}
