% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{vismclDfDraw}
\alias{vismclDfDraw}
\title{vismclDfDraw}
\usage{
vismclDfDraw(input_df, showName = FALSE, Clusternames = NULL,
  color = "grey", HighlightByName = NULL, HighlightFirstN = NULL,
  HighlightColor = "chartreuse1")
}
\arguments{
\item{input_df}{dataframe created by CreateDataFrame}

\item{showName}{(optional)if shows the name of clusters,default is FALSE}

\item{Clusternames}{(optional) a vector for user-defined clusternames,
order should be the same as the lines go.}

\item{color}{(optional)A color code:color fill the circle,default is grey}

\item{HighlightByName}{(optional) a vector of cluster names which user wish to highlight.}

\item{HighlightFirstN}{(optional) an integer:highlight the first n largest cluster.}

\item{HighlightColor}{(optional) A color code: color used for highlight, default is green}
}
\value{
a circle packing graph representing the information
}
\description{
plot circle packing graph with data frame.
}
\examples{
x <- data.frame("Name" = 1:2, "Element" = c("a","b"), "Area" = c(1,1))
vismclDfDraw(x,showName = TRUE,Clusternames=c("1","2"),
color="white",HighlightByName="1",HighlightFirstN=1,HighlightColor="red")
}
