% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismclDFDraw.R
\name{AddToList}
\alias{AddToList}
\title{AddToList}
\usage{
AddToList(filepath)
}
\arguments{
\item{filepath}{the filepath of output of MCL.}
}
\value{
A nested list.
}
\description{
add the text file from output of MCL to a nested list.
}
