% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{CreatDataFrame}
\alias{CreatDataFrame}
\title{CreatDataFrame}
\usage{
CreatDataFrame(filepath)
}
\arguments{
\item{filepath}{filepath of the output of MCL}
}
\value{
a dataframe used for plotting.
}
\description{
creat dataframe used for plotting from nested list
}
\details{
Details.
}
