% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{vismcl}
\alias{vismcl}
\title{Title: plot circle packing graph}
\usage{
vismcl(filepath, showName = FALSE, color = "grey")
}
\arguments{
\item{filepath}{the filepath of the output file of MCL}

\item{showName}{if shows the name of clusters,default is FALSE}

\item{color}{A color code:color fill the circle,default is grey#'}
}
\value{
a circle packing graph representing the information
}
\description{
plot circle packing graph with MCL output file,with interactive display. The input file should not have any head or tail content.
}
\examples{
vismcl("MCL_example_output",TRUE,"red")
vismcl("MCL_example_output",FALSE)
}
