% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl(L).R
\name{vismcl}
\alias{vismcl}
\title{vismcl}
\usage{
vismcl(filepath, showName = FALSE, Clusternames = NULL,
  color = "grey", HighlightByName = NULL, HighlightFirstN = 0,
  HighlightColor = "chartreuse1")
}
\arguments{
\item{filepath}{the filepath of the output file of MCL}

\item{showName}{(optional)if shows the name of clusters,default is FALSE}

\item{Clusternames}{(optional) a vector for user-defined clusternames,
order should be the same as the lines go.}

\item{color}{(optional)A color code:color fill the circle,default is grey}

\item{HighlightByName}{(optional) a vector of cluster names which user wish to highlight.}

\item{HighlightFirstN}{(optional) an integer:highlight the first n largest cluster.}

\item{HighlightColor}{(optional) A color code: color used for highlight, default is green}
}
\value{
a circle packing graph representing the information and print the dataframe in the console
}
\description{
plot circle packing graph with given filepath,recommand vismclL for very large input data
}
\examples{
cat("a aa\\n b bbb b\\n",file = "example")
vismcl("example",showName = TRUE,Clusternames=c("1","2"),color="white",HighlightByName="1",
       HighlightFirstN=1,HighlightColor="red")
vismcl("example")
vismcl("example",showName = TRUE)
}
