% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vismcl.R
\name{vismclL}
\alias{vismclL}
\title{vismclL}
\usage{
vismclL(filepath, filter, showName = FALSE, Clusternames = NULL,
  color = "grey", HighlightByName = NULL, HighlightFirstN = NULL,
  HighlightColor = "chartreuse1")
}
\arguments{
\item{filepath}{the filepath of the output file of MCL}

\item{filter}{the first n largest clusters will be plotted. if there are more than one
cluster with the cutoff size,vismclL will print all of them.}

\item{showName}{(optional)if shows the name of clusters,default is FALSE}

\item{Clusternames}{(optional) a vector for user-defined clusternames,
order should be the same as the lines go.}

\item{color}{(optional)A color code:color fill the circle,default is grey}

\item{HighlightByName}{(optional) a vector of cluster names which user wish to highlight.}

\item{HighlightFirstN}{(optional) an integer:highlight the first n largest cluster.}

\item{HighlightColor}{(optional) A color code: color used for highlight, default is green}
}
\value{
a circle packing graph representing the information
}
\description{
plot circle packing graph with given filepath,
with an extra filter attribute to only print the largest n clusters
}
\examples{
cat("a aa\\n b bbb b\\n",file = "example")
vismclL("example",filter=1,showName = TRUE,Clusternames=c("1","2"),color="white",HighlightByName="1",
       HighlightFirstN=1,HighlightColor="red")
}
